payment-recovery-system-automation/requirements.txt
# Python Automation Service Dependencies

# HTTP Client for Java API
requests==2.31.0
urllib3==2.1.0

# Task Scheduling
APScheduler==3.10.4
celery==5.3.4

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Email Support
email-validator==2.1.0

# Logging
structlog==24.1.0

# Date/Time Utilities
python-dateutil==2.8.2

# Type Hints
typing-extensions==4.9.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

